Further separate module run, directory creation, and disk write
