+++ multiqc/core/__init__.py
@@ -1,0 +1,14 @@
+from .exec_modules import exec_modules
+from .file_search import file_search
+from .init_config import init_config
+from .write_results import write_results
+from .exceptions import RunResult, RunError
+
+__all__ = [
+    "exec_modules",
+    "file_search",
+    "init_config",
+    "write_results",
+    "RunResult",
+    "RunError",
+]
