+++ multiqc/core/exceptions.py
@@ -1,5 +1,4 @@
 import dataclasses
-import os
 
 from multiqc.utils import config, report
 
@@ -22,7 +21,7 @@ class RunResult:
     config = config
 
 
-class _RunError(Exception):
+class RunError(Exception):
     """
     Used internally in `run` to pass errors from sub-steps.
     """
@@ -30,11 +29,3 @@ class _RunError(Exception):
     def __init__(self, message: str = "", sys_exit_code: int = 1):
         self.message = message
         self.sys_exit_code = sys_exit_code
-
-
-def _data_tmp_dir() -> str:
-    return os.path.join(report.tmp_dir, "multiqc_data")
-
-
-def _plots_tmp_dir() -> str:
-    return os.path.join(report.tmp_dir, "multiqc_plots")
