+++ multiqc/modules/base_module.py
@@ -15,7 +15,7 @@ from collections import defaultdict
 import markdown
 
 from multiqc.plots.plotly.plot import Plot
-from multiqc.utils import config, report, software_versions, util_functions
+from multiqc.utils import config, report, software_versions
 
 logger = logging.getLogger(__name__)
 
@@ -294,7 +294,6 @@ class BaseMultiqcModule:
         description = description.strip()
         comment = comment.strip()
         helptext = helptext.strip()
-        plot_html = plot.add_to_report(report) if isinstance(plot, Plot) else (plot or "")
 
         self.sections.append(
             {
@@ -304,14 +303,9 @@ class BaseMultiqcModule:
                 "comment": comment,
                 "helptext": helptext,
                 "content_before_plot": content_before_plot,
-                "plot": plot_html,
+                "plot": plot,
                 "content": content,
-                "print_section": any(
-                    [
-                        n is not None and len(n) > 0
-                        for n in [description, comment, helptext, content_before_plot, plot_html, content]
-                    ]
-                ),
+                "print_section": any([description, comment, helptext, content_before_plot, plot, content]),
             }
         )
 
@@ -632,7 +626,7 @@ class BaseMultiqcModule:
         # Keep also in the module instance, so it's possible to map back data to specific module
         self.__saved_raw_data[fn] = data
 
-        util_functions.write_data_file(data, fn, sort_cols, data_format)
+        report.write_data_file(data, fn, sort_cols, data_format)
 
     ##################################################
     #### DEPRECATED FORWARDERS
